feat: Faster reloading
Instead of letting Bubble Tea shut down entirely, immediately `exec()`.
This means less delay and less flashing if the program is on the
alternate screen, but risks leaving the terminal in an odd state.

We address that by manually capturing the terminal state and restoring
it just before `exec()`, which is faster than a complete proper
shutdown. However, this still assumes that the input is `os.Stdin` and a
TTY, which Bubble Tea does not assume. This is a limitation of Resteep
for now.